Restore SearchParametersHashUpdatedHandler.cs
